Refactor of Retriever Report
1. support double of error info
2. add runtime protection for Fraction with 0 denumerator
3. refactor of SteadyClock

+++ objc/source/interface/statictics/WCTError.h
@@ -99,7 +99,7 @@ typedef NS_ENUM(NSUInteger, WCTErrorCode) {
 
 @interface WCTError (ExtendedCode)
 
-- (NSInteger)extendedCode;
+- (int)extendedCode;
 
 @end
